+++ NEWS.md
@@ -1,5 +1,9 @@
 # ggnewscale (development version)
 
+## Bugfixes
+
+-   `new_scale()` respects `override.aes` arguments (#98, thanks @jgabry)
+
 # ggnewscale 0.4.6
 
 ## Bugfixes